Practice p[roblems

+++ sqrtOfNum.py
@@ -2,7 +2,7 @@ import math
 
 
 print("This programme is to find the square root of any number")
-num=int(input("enter the number: "))
+num=float(input("enter the number: "))
 sqrt=num**0.5
 print(math.sqrt(num))  #by using library
 print("sqrt of",num,"is",sqrt)
